Add ge_basis and refactor siemens_basis in a spherical harmonics basis file
